+++ templates/show_shout_row.inc.php
@@ -24,7 +24,7 @@
     <td class="cel_username"><?php echo $client->f_link; ?></td>
     <td class="cel_sticky"><?php echo $libitem->sticky; ?></td>
     <td class="cel_comment"><?php echo scrub_out($libitem->text); ?></td>
-    <td class="cel_date"><?php echo $libitem->date; ?></td>
+    <td class="cel_date"><?php echo $libitem->f_date; ?></td>
     <td class="cel_action">
         <a href="<?php echo $web_path; ?>/admin/shout.php?action=show_edit&amp;shout_id=<?php echo $libitem->id; ?>">
             <?php echo UI::get_icon('edit', T_('Edit')); ?>